test(engine): add reproduction of #6152
This test reproduces the problem described in github issue #6152. A
multi instance sub process containing an embedded non-interrupting event
sub process which is cancelled. The termination of the embedded
subprocess should inform the multi-instance sub process that it has
terminated, but there seems to be a mismatch in the selected processor.