chore(engine): refactor logic to trigger an event
This commit achieves an intermediate state in which the JobThrowErrorProcessor does not modify state.
It is an intermediate state, because the JobThrowErrorProcessor still writes the EVENT_OCCURRED record,
whilst the ZEP mandates that instead the element shall be terminated and the termination of the element
shall trigger the event element. A new issue was created to address this follow up work.

This commit also expands the random workflow generation to now also generate
boundary error events for service tasks.

There are two behavior changes that break with the goal of a simple refactoring:
- the trigger event method is now called with the jobKey as second parameter.
This was suggested and approved by Phiipp. According to his info the parameter
is not used later an, and this branch of code might not live very long (see above)
- there is now a magic ID to indicate that no catch event could be found.
This is to avoid looking for the catch event again and again. Currently, the lookup
still has to be repeated between processor and applier. After the follow up work is done,
most likely the applier won't need to do the lookup anymore.